Remove unused webmention draft entry, add date to observable entry.
For some reason, pelican is adding a very weird date to certain drafts that have no date, and this is causing compilcation problems.  I don't know why this is happening *now*, but it is, so I'm removing one of the offending entries, which had no content anyway, and I'm adding a date to the other.